requirements.txt
asgiref==3.9.1
channels==4.3.1
channels_redis==4.3.0
coverage==7.9.2
Django==5.2.4
django-cors-headers==4.7.0
django-filter==25.1
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
msgpack==1.1.1
PyJWT==2.9.0
redis==6.4.0
sqlparse==0.5.3
